equip and unequip work, add stats and tex next

+++ Necromancy.Server/Packet/Area/send_item_equip.cs
@@ -19,7 +19,7 @@ namespace Necromancy.Server.Packet.Area
             ItemZoneType zone = (ItemZoneType) packet.Data.ReadByte();
             byte bag = packet.Data.ReadByte();
             short slot = packet.Data.ReadInt16();
-            ItemEquipSlots equipSlot = (ItemEquipSlots) packet.Data.ReadByte();
+            ItemEquipSlots equipSlot = (ItemEquipSlots) packet.Data.ReadInt32();
             
             Logger.Debug($"storageType:{zone} bagId:{bag} bagSlotIndex:{slot} equipBit:{equipSlot}");
 
@@ -37,37 +37,6 @@ namespace Necromancy.Server.Packet.Area
 
             RecvItemEquip recvItemEquip = new RecvItemEquip(client, error);
             Router.Send(recvItemEquip);
-
-            //short PhysAttack = 0;
-            //short MagAttack = 0;
-            //short PhysDef = 0;
-            //short MagDef = 0;
-            //TODO
-            //foreach (InventoryItem inventoryItem2 in client.Character.Inventory._equippedItems.Values)
-            //{
-            //    if ((int)inventoryItem2.CurrentEquipmentSlotType < 3)
-            //    {
-            //        PhysAttack += (short)inventoryItem2.Item.Physical;
-            //        MagAttack += (short)inventoryItem2.Item.Magical;
-            //    }
-            //    else
-            //    {
-            //        PhysDef += (short)inventoryItem2.Item.Physical;
-            //        MagDef += (short)inventoryItem2.Item.Magical;
-            //    }
-            //}
-
-            //res = BufferProvider.Provide();
-            //res.WriteInt16((short)client.Character.Strength); //base Phys Attack
-            //res.WriteInt16((short)client.Character.intelligence); //base Mag attack
-            //res.WriteInt16((short)client.Character.dexterity); //base Phys Def
-            //res.WriteInt16((short)client.Character.piety); //base Mag Def
-
-            //res.WriteInt16(PhysAttack); //Equip Bonus Phys attack
-            //res.WriteInt16(MagAttack); //Equip Bonus Mag Attack
-            //res.WriteInt16(PhysDef); //Equip bonus Phys Def
-            //res.WriteInt16(MagDef); //Equip bonus Mag Def
-            //Router.Send(client, (ushort)AreaPacketId.recv_chara_update_battle_base_param, res, ServerType.Area);
         }
     }
 }
